fixed options within attributes

+++ src/Models/Loyalty/Receptions/CreditReception.php
@@ -63,10 +63,10 @@ class CreditReception
      * @param Shop $shop
      * @param ContactIdentifier|null $contactIdentifier
      * @param DateTime $createdAt
-     * @param int|null $unitValue
+     * @param float|null $unitValue
      * @param Unit|null $unit
      */
-    public function __construct(string $type, int $credits, string $uuid, Contact $contact, Shop $shop, ?ContactIdentifier $contactIdentifier, DateTime $createdAt, ?int $unitValue = null, ?Unit $unit = null)
+    public function __construct(string $type, int $credits, string $uuid, Contact $contact, Shop $shop, ?ContactIdentifier $contactIdentifier, DateTime $createdAt, ?float $unitValue = null, ?Unit $unit = null)
     {
         $this->type = $type;
         $this->credits = $credits;
@@ -136,9 +136,9 @@ class CreditReception
     }
 
     /**
-     * @return int|null
+     * @return float|null
      */
-    public function getUnitValue(): ?int
+    public function getUnitValue(): ?float
     {
         return $this->unitValue;
     }
